[frontend] refactor folder structure
