+++ bankme-frontend/src/components/payable/update-payable.tsx
@@ -1,4 +1,3 @@
-"use client";
 import { z } from "zod";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
